refactor: move out u-saint application and models as separate feature
