+++ src/session/mod.rs
@@ -1,18 +1,22 @@
-use std::{borrow::BorrowMut, ops::Deref};
+use std::{borrow::BorrowMut, ops::Deref, sync::Arc};
 
 use reqwest::{
     cookie::{CookieStore, Jar},
     header::{HeaderValue, COOKIE, HOST, SET_COOKIE},
+    Client,
 };
 use url::Url;
 
 use crate::{
-    utils::{default_header, obtain_ssu_sso_token, DEFAULT_USER_AGENT},
-    webdynpro::error::{ClientError, WebDynproError}, error::RusaintError,
+    error::{RusaintError, SsuSsoError},
+    utils::{default_header, DEFAULT_USER_AGENT},
+    webdynpro::error::{ClientError, WebDynproError},
 };
 
 const SSU_USAINT_PORTAL_URL: &str = "https://saint.ssu.ac.kr/irj/portal";
 const SSU_USAINT_SSO_URL: &str = "https://saint.ssu.ac.kr/webSSO/sso.jsp";
+const SMARTID_LOGIN_URL: &str = "https://smartid.ssu.ac.kr/Symtra_sso/smln.asp";
+const SMARTID_LOGIN_FORM_REQUEST_URL: &str = "https://smartid.ssu.ac.kr/Symtra_sso/smln_pcs.asp";
 
 #[derive(Debug, Default)]
 pub struct USaintSession(Jar);
@@ -118,13 +122,66 @@ impl USaintSession {
 
     pub async fn with_password(id: &str, password: &str) -> Result<USaintSession, RusaintError> {
         let token = obtain_ssu_sso_token(id, password).await?;
-        Ok(Self::with_token(id, &token).await.or_else(|e| Err(WebDynproError::ClientError(e)))?)
+        Ok(Self::with_token(id, &token)
+            .await
+            .or_else(|e| Err(WebDynproError::ClientError(e)))?)
     }
+}
+
+pub async fn obtain_ssu_sso_token(id: &str, password: &str) -> Result<String, SsuSsoError> {
+    let jar: Arc<Jar> = Arc::new(Jar::default());
+    let client = Client::builder()
+        .cookie_provider(jar)
+        .cookie_store(true)
+        .user_agent(DEFAULT_USER_AGENT)
+        .build()
+        .unwrap();
+    let body = client
+        .get(SMARTID_LOGIN_URL)
+        .headers(default_header())
+        .send()
+        .await?
+        .text()
+        .await?;
+    let document = scraper::Html::parse_document(&body);
+    let in_tp_bit_selector = scraper::Selector::parse(r#"input[name="in_tp_bit"]"#).unwrap();
+    let rqst_caus_cd_selector = scraper::Selector::parse(r#"input[name="rqst_caus_cd"]"#).unwrap();
+    let in_tp_bit = document
+        .select(&in_tp_bit_selector)
+        .next()
+        .ok_or(SsuSsoError::CantLoadForm)?
+        .value()
+        .attr("value")
+        .ok_or(SsuSsoError::CantLoadForm)?;
+    let rqst_caus_cd = document
+        .select(&rqst_caus_cd_selector)
+        .next()
+        .ok_or(SsuSsoError::CantLoadForm)?
+        .value()
+        .attr("value")
+        .ok_or(SsuSsoError::CantLoadForm)?;
+    let params = [
+        ("in_tp_bit", in_tp_bit),
+        ("rqst_caus_cd", rqst_caus_cd),
+        ("userid", id),
+        ("pwd", password),
+    ];
+    let res = client
+        .post(SMARTID_LOGIN_FORM_REQUEST_URL)
+        .headers(default_header())
+        .form(&params)
+        .send()
+        .await?;
+    let cookie_token = res
+        .cookies()
+        .find(|cookie| cookie.name() == "sToken" && !cookie.value().is_empty())
+        .ok_or(SsuSsoError::CantFindToken)?;
+    Ok(cookie_token.value().to_string())
 }
 
 #[cfg(test)]
 mod test {
-    use crate::session::USaintSession;
+    use crate::session::{obtain_ssu_sso_token, USaintSession};
     use dotenv::dotenv;
 
     #[tokio::test]
@@ -134,5 +191,14 @@ mod test {
         let password = std::env::var("SSO_PASSWORD").unwrap();
         let session = USaintSession::with_password(&id, &password).await.unwrap();
         println!("{:?}", session);
+    }
+
+    #[tokio::test]
+    async fn test_obtain_sso_token() {
+        dotenv().ok();
+        let id = std::env::var("SSO_ID").unwrap();
+        let password = std::env::var("SSO_PASSWORD").unwrap();
+        let token = obtain_ssu_sso_token(&id, &password).await.unwrap();
+        assert_ne!("", token);
     }
 }
